+++ fvdb/src/detail/ops/FillFromGrid.cu
@@ -55,7 +55,7 @@ fillToGridCPU(const GridBatchImpl::Accessor<nanovdb::ValueOnIndex> &fromGridHand
               const torch::TensorAccessor<ScalarType, 2> fromFeatures,
               torch::TensorAccessor<ScalarType, 2> toFeatures,
               bool isContiguous) {
-    for (size_t bi = 0; bi < fromGridHandle.batchSize(); bi += 1) {
+    for (int64_t bi = 0; bi < fromGridHandle.batchSize(); bi += 1) {
         const nanovdb::OnIndexGrid *fromGrid = fromGridHandle.grid(bi);
         const nanovdb::OnIndexGrid *toGrid   = toGridHandle.grid(bi);
